requirements.txt
Babel==2.4.0
Django==1.10.6
django-phonenumber-field==1.2.0
oauthlib==2.0.2
olefile==0.44
phonenumberslite==8.3.3
Pillow==4.0.0
psycopg2==2.7.1
PyJWT==1.4.2
python-openid==2.2.5
python-social-auth==0.2.21
pytz==2016.10
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-core==1.2.0

